# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistem Pametne Kuće API",
        "description": "API dokumentacija za upravljanje uređajima, korisnicima i ulogama (Admin, Običan, Dete).",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Lokalni razvojni server"
        }
    ],
    "paths": {
...
        "/api/stanje-uredjaja": {
            "get": {
                "tags": [
                    "Stanja Uređaja (Komande)"
                ],
                "summary": "Istorija i trenutna stanja svih uređaja",
                "operationId": "a091b7a56dc3a5457571138bedeb2b5f",
                "responses": {
                    "200": {
                        "description": "Uspešno učitana stanja"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Stanja Uređaja (Komande)"
                ],
                "summary": "Dodaj uređaj u sobu (kreiraj početno stanje)",
                "operationId": "d7f1d9d7e0b8d8ee0ea086b33f26b7cc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nazivUredjaja": {
                                        "type": "string",
                                        "example": "Klima Dnevna"
                                    },
                                    "ukljucen": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "podesavanja": {
                                        "type": "object",
                                        "example": {
                                            "temp": 24,
                                            "mod": "hladjenje"
                                        }
                                    },
                                    "uredjaj_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "soba_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Stanje uspešno kreirano"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/stanje-uredjaja/{id}": {
            "put": {
                "tags": [
                    "Stanja Uređaja (Komande)"
                ],
                "summary": "Izvrši komandu (Promeni stanje/podešavanja uređaja)",
                "description": "Koristi se za paljenje/gašenje ili promenu parametara poput temperature.",
                "operationId": "9232ebed6fcbf12e2d2ad4cbe3c3ba96",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ukljucen": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "podesavanja": {
                                        "type": "object",
                                        "example": {
                                            "temp": 22
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Komanda uspešno izvršena"
                    },
                    "404": {
                        "description": "Uređaj nije pronađen"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Stanja Uređaja (Komande)"
                ],
                "summary": "Ukloni uređaj iz sobe (briše stanje)",
                "operationId": "d6520bd1f066c28c71aba655fb1b611a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stanje obrisano"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/uredjaji": {
            "get": {
                "tags": [
                    "Katalog Uređaja"
                ],
                "summary": "Vraća listu svih dostupnih uređaja u katalogu",
                "operationId": "56c56d452fe3b9b82d0c9917cb27a36e",
                "responses": {
                    "200": {
                        "description": "Uspešno učitana lista uređaja"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Katalog Uređaja"
                ],
                "summary": "Dodavanje novog tipa uređaja u katalog",
                "operationId": "3676b73f7e271be294b0021c0a9484c4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "marka": {
                                        "type": "string",
                                        "example": "Samsung"
                                    },
                                    "model": {
                                        "type": "string",
                                        "example": "WindFree Gen2"
                                    },
                                    "tipUredjaja": {
                                        "type": "string",
                                        "example": "Klima"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Uređaj uspešno dodat u katalog"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/uredjaji/{id}": {
            "get": {
                "tags": [
                    "Katalog Uređaja"
                ],
                "summary": "Detalji o uređaju sa istorijom stanja",
                "operationId": "b5afa3ae38c22e3134ec344e58d58b16",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno učitani detalji uređaja"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Katalog Uređaja"
                ],
                "summary": "Izmena informacija o uređaju u katalogu",
                "operationId": "305b14da41234849b86579ac3a7ff8ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "marka": {
                                        "type": "string"
                                    },
                                    "model": {
                                        "type": "string"
                                    },
                                    "tipUredjaja": {
                                        "type": "string",
                                        "enum": [
                                            "Klima",
                                            "Svetlo",
                                            "Grejalica"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Uređaj uspešno ažuriran"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Katalog Uređaja"
                ],
                "summary": "Brisanje uređaja iz kataloga",
                "operationId": "752467abc2a46507513a97bce7936caf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uređaj obrisan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Ovde unesite access_token dobijen nakon logovanja.",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}